[FIX] verify if file existes before delete it

+++ lib/terminal_commands.dart
@@ -3,7 +3,9 @@ import 'dart:io';
 class TerminalCommand {
   static Future<void> deleteFile(String filePath) async {
     if (Platform.isWindows) {
-      Directory(filePath).deleteSync(recursive: true);
+      if (Directory(filePath).existsSync()) {
+        Directory(filePath).deleteSync(recursive: true);
+      }
     } else {
       Process.runSync('rm', ['-r', filePath]);
     }
